+++ basex-core/src/test/java/org/basex/util/MinHeapTest.java
@@ -28,7 +28,7 @@ public final class MinHeapTest {
 
     int i = 0;
     while(!heap.isEmpty()) {
-      assertEquals(Integer.valueOf(i++), heap.removeMin());
+      assertEquals(i++, heap.removeMin());
       heap.verify();
     }
     assertEquals(1000, i);
@@ -44,7 +44,7 @@ public final class MinHeapTest {
 
     int i = 0;
     while(!heap.isEmpty()) {
-      assertEquals(i++, heap.removeMin().intValue());
+      assertEquals(i++, heap.removeMin());
       heap.verify();
     }
     assertEquals(1000, i);
